+++ src/z3c/autoinclude/dependency.py
@@ -7,6 +7,10 @@ from pkg_resources import get_distribution
 from z3c.autoinclude.utils import DistributionManager
 from z3c.autoinclude.utils import ZCMLInfo
 
+
+logger = logging.getLogger("z3c.autoinclude")
+
+
 class DependencyFinder(DistributionManager):
 
     def includableInfo(self, zcml_to_look_for):
@@ -25,12 +29,17 @@ class DependencyFinder(DistributionManager):
                 try:
                     module = resolve(dotted_name)
                 except ImportError as exc:
-                    logging.getLogger("z3c.autoinclude").warn(
+                    logger.warning(
                         "resolve(%r) raised import error: %s" % (dotted_name, exc))
+                    continue
+                module_file = getattr(module, '__file__', None)
+                if module_file is None:
+                    logger.warning(
+                        "%r has no __file__ attribute" % dotted_name)
                     continue
                 for candidate in zcml_to_look_for:
                     candidate_path = os.path.join(
-                        os.path.dirname(module.__file__), candidate)
+                        os.path.dirname(module_file), candidate)
                     if os.path.isfile(candidate_path):
                         result[candidate].append(dotted_name)
         return result